+++ ors-engine/src/main/java/org/heigit/ors/routing/graphhopper/extensions/util/parsers/AccessRestrictionsParser.java
@@ -14,44 +14,21 @@ import java.util.*;
 
 
 public class AccessRestrictionsParser implements TagParser {
-    private final IntEncodedValue accessRestrictionEnc;
-
     private static final String VAL_ACCESS = "access";
     private static final String VAL_MOTOR_VEHICLE = "motor_vehicle";
     private static final String VAL_MOTORCAR = "motorcar";
     private static final String VAL_VEHICLE = "vehicle";
     private static final String VAL_MOTORCYCLE = "motorcycle";
+    private static final String VAL_HGV = "hgv";
     private static final String VAL_BICYCLE = "bicycle";
     private static final String VAL_FOOT = "foot";
-    private final List<String> accessRestrictedTags = new ArrayList<>(5);
-    private final List<String> motorCarTags = new ArrayList<>(5);
-    private final List<String> motorCycleTags = new ArrayList<>(5);
-    private final Set<String> restrictedValues = new HashSet<>(5);
-    private final Set<String> permissiveValues = new HashSet<>(5);
-
+    private static final List<String> motorcarTags = Arrays.asList(VAL_MOTORCAR, VAL_MOTOR_VEHICLE, VAL_VEHICLE, VAL_ACCESS);
+    private static final List<String> motorcycleTags = Arrays.asList(VAL_MOTORCYCLE, VAL_MOTOR_VEHICLE, VAL_VEHICLE, VAL_ACCESS);
+    private static final List<String> hgvTags = Arrays.asList(VAL_HGV, VAL_MOTOR_VEHICLE, VAL_VEHICLE, VAL_ACCESS);
+    private static final List<String> bicycleTags = Arrays.asList(VAL_BICYCLE, VAL_VEHICLE, VAL_ACCESS);
+    private static final List<String> footTags = Arrays.asList(VAL_FOOT, VAL_ACCESS);
+    private final IntEncodedValue accessRestrictionEnc;
     private final int profileType;
-
-    public void initTags() {
-        accessRestrictedTags.addAll(Arrays.asList(VAL_MOTORCAR, VAL_MOTOR_VEHICLE, VAL_VEHICLE, VAL_ACCESS, VAL_BICYCLE, VAL_FOOT));
-        motorCarTags.addAll(Arrays.asList(VAL_MOTORCAR, VAL_MOTOR_VEHICLE));
-        motorCycleTags.addAll(Arrays.asList(VAL_MOTORCYCLE, VAL_MOTOR_VEHICLE));
-
-        restrictedValues.add("private");
-        restrictedValues.add("no");
-        restrictedValues.add("restricted");
-        restrictedValues.add("military");
-        restrictedValues.add("destination");
-        restrictedValues.add("customers");
-        restrictedValues.add("emergency");
-        restrictedValues.add("permissive");
-        restrictedValues.add("delivery");
-        restrictedValues.add("permit");
-
-        permissiveValues.add("yes");
-        permissiveValues.add("designated");
-        permissiveValues.add("official");
-    }
-
 
     public AccessRestrictionsParser(int profileType) {
         this(AccessRestriction.create(), profileType);
@@ -60,7 +37,6 @@ public class AccessRestrictionsParser implements TagParser {
     public AccessRestrictionsParser(IntEncodedValue accessRestrictionEnc, int profileType) {
         this.accessRestrictionEnc = accessRestrictionEnc;
         this.profileType = profileType;
-        initTags();
     }
 
     @Override
@@ -76,39 +52,26 @@ public class AccessRestrictionsParser implements TagParser {
     }
 
     public int processWay(ReaderWay way) {
-        //TODO: modify the following logic to process access restriction tags from the most specific to the least
-        //      specific, e.g. motorcar > motor_vehicle > vehicle > access via a call to way.getFirstPriorityTag
-        //      see https://github.com/GIScience/openrouteservice/issues/2275
-
-        if (!way.hasTag(accessRestrictedTags, restrictedValues)) {
-            return 0;
-        }
-
-        if (RoutingProfileType.isDriving(profileType)) {
-            return processAccess(way, motorCarTags);
+        if (profileType == RoutingProfileType.DRIVING_CAR) {
+            return getRestrictionType(way, motorcarTags);
         }
         if (profileType == RoutingProfileType.DRIVING_MOTORCYCLE) {
-            return processAccess(way, motorCycleTags);
+            return getRestrictionType(way, motorcycleTags);
+        }
+        if (RoutingProfileType.isHeavyVehicle(profileType)) {
+            return getRestrictionType(way, hgvTags);
         }
         if (RoutingProfileType.isCycling(profileType)) {
-            return processAccess(way, VAL_BICYCLE);
+            return getRestrictionType(way, bicycleTags);
         }
         if (RoutingProfileType.isPedestrian(profileType)) {
-            return processAccess(way, VAL_FOOT);
+            return getRestrictionType(way, footTags);
         }
         if (profileType == RoutingProfileType.UNKNOWN) {
-            return processAccess(way, VAL_ACCESS);
+            return getRestrictionType(way, Collections.singletonList(VAL_ACCESS));
         }
 
         return 0;
-    }
-
-    private int processAccess(ReaderWay way, String tag) {
-        return processAccess(way, Collections.singletonList(tag));
-    }
-
-    private int processAccess(ReaderWay way, List<String> tags) {
-        return isAccessAllowed(way, tags) ? 0 : getRestrictionType(way, tags);
     }
 
     /**
@@ -121,15 +84,8 @@ public class AccessRestrictionsParser implements TagParser {
     private int getRestrictionType(ReaderWay way, List<String> tags) {
         int res = 0;
 
-        String tagValue = way.getTag(VAL_ACCESS);
-        if (tagValue != null)
-            res = updateRestriction(res, tagValue);
-
-        if (tags != null) {
-            for (String key : tags) {
-                tagValue = way.getTag(key);
-                res = updateRestriction(res, tagValue);
-            }
+        for (String value : way.getFirstPriorityTagValues(tags)) {
+            res = updateRestriction(res, value);
         }
 
         return res;
@@ -145,16 +101,4 @@ public class AccessRestrictionsParser implements TagParser {
     private int updateRestriction(int encodedRestrictions, String restrictionValue) {
         return encodedRestrictions | AccessRestrictionType.getFromString(restrictionValue);
     }
-
-    /**
-     * Check if access is allowed on the way. e.g. it would check if motor_car=yes/permissive/destination etc. is set
-     *
-     * @param way      The OSM way to be checked
-     * @param tagNames The tags (keys) to be checked
-     * @return Whether access is allowed on the way
-     */
-    private boolean isAccessAllowed(ReaderWay way, List<String> tagNames) {
-        return way.hasTag(tagNames, permissiveValues);
-    }
-
 }