+++ ors-engine/src/test/java/org/heigit/ors/routing/graphhopper/extensions/util/parsers/AccessRestrictionParserTest.java
@@ -25,9 +25,8 @@ class AccessRestrictionParserTest {
         intsRef = em.createEdgeFlags();
     }
 
-    @Test
-    void testAccessTypes() {
-        initParser(RoutingProfileType.UNKNOWN);
+    private void assertRestrictionsForProfile(int profileType) {
+        initParser(profileType);
         ReaderWay way = new ReaderWay(1);
 
         parser.handleWayTags(intsRef, way, false, relFlags);
@@ -63,6 +62,16 @@ class AccessRestrictionParserTest {
     }
 
     @Test
+    void testAccessTypesCar() {
+        assertRestrictionsForProfile(RoutingProfileType.DRIVING_CAR);
+    }
+
+    @Test
+    void testAccessCycling() {
+        assertRestrictionsForProfile(RoutingProfileType.CYCLING_REGULAR);
+    }
+
+    @Test
     void testCarWayCreation() {
         initParser(RoutingProfileType.DRIVING_CAR);
 
@@ -73,11 +82,103 @@ class AccessRestrictionParserTest {
 
         way.setTag("motorcar", "destination");
         parser.handleWayTags(intsRef, way, false, relFlags);
-        assertEquals(AccessRestrictionType.NO | AccessRestrictionType.DESTINATION, accessRestrictionEnc.getInt(false, intsRef));
+        assertEquals(AccessRestrictionType.DESTINATION, accessRestrictionEnc.getInt(false, intsRef));
 
         way.setTag("motorcar", "yes");
         parser.handleWayTags(intsRef, way, false, relFlags);
         assertEquals(AccessRestrictionType.NONE, accessRestrictionEnc.getInt(false, intsRef));
+    }
+
+    @Test
+    void testVehicleDestination() {
+        initParser(RoutingProfileType.DRIVING_CAR);
+
+        ReaderWay way = new ReaderWay(1);
+
+        way.setTag("vehicle", "destination");
+        parser.handleWayTags(intsRef, way, false, relFlags);
+        assertEquals(AccessRestrictionType.DESTINATION, accessRestrictionEnc.getInt(false, intsRef));
+    }
+
+    @Test
+    void testHgvDestination() {
+        initParser(RoutingProfileType.DRIVING_HGV);
+
+        ReaderWay way = new ReaderWay(1);
+
+        way.setTag("hgv", "destination");
+        parser.handleWayTags(intsRef, way, false, relFlags);
+        assertEquals(AccessRestrictionType.DESTINATION, accessRestrictionEnc.getInt(false, intsRef));
+    }
+
+    @Test
+    void testAccessNoVehicleDestination() {
+        initParser(RoutingProfileType.DRIVING_CAR);
+
+        ReaderWay way = new ReaderWay(1);
+
+        way.setTag("access", "no");
+        way.setTag("vehicle", "destination");
+        parser.handleWayTags(intsRef, way, false, relFlags);
+        assertEquals(AccessRestrictionType.DESTINATION, accessRestrictionEnc.getInt(false, intsRef));
+    }
+
+    @Test
+    void testMotorcarDestinationMotorvehicleYes() {
+        initParser(RoutingProfileType.DRIVING_CAR);
+
+        ReaderWay way = new ReaderWay(1);
+
+        way.setTag("motorcar", "destination");
+        way.setTag("motor_vehicle", "yes");
+        parser.handleWayTags(intsRef, way, false, relFlags);
+        assertEquals(AccessRestrictionType.DESTINATION, accessRestrictionEnc.getInt(false, intsRef));
+    }
+
+    @Test
+    void testMotorcarYesVehicleDestination() {
+        initParser(RoutingProfileType.DRIVING_CAR);
+
+        ReaderWay way = new ReaderWay(1);
+
+        way.setTag("motorcar", "yes");
+        way.setTag("vehicle", "destination");
+        parser.handleWayTags(intsRef, way, false, relFlags);
+        assertEquals(AccessRestrictionType.NONE, accessRestrictionEnc.getInt(false, intsRef));
+    }
+
+    @Test
+    void testMotorcarDestinationAccessNo() {
+        initParser(RoutingProfileType.DRIVING_CAR);
+
+        ReaderWay way = new ReaderWay(1);
+
+        way.setTag("motorcar", "destination");
+        way.setTag("access", "no");
+        parser.handleWayTags(intsRef, way, false, relFlags);
+        assertEquals(AccessRestrictionType.DESTINATION, accessRestrictionEnc.getInt(false, intsRef));
+    }
+
+    @Test
+    void testAccessPrivateCustomers() {
+        initParser(RoutingProfileType.DRIVING_CAR);
+
+        ReaderWay way = new ReaderWay(1);
+
+        way.setTag("access", "private;customers");
+        parser.handleWayTags(intsRef, way, false, relFlags);
+        assertEquals(AccessRestrictionType.PRIVATE | AccessRestrictionType.CUSTOMERS, accessRestrictionEnc.getInt(false, intsRef));
+    }
+
+    @Test
+    void testMotorcarPrivateCustomers() {
+        initParser(RoutingProfileType.DRIVING_CAR);
+
+        ReaderWay way = new ReaderWay(1);
+
+        way.setTag("motorcar", "private;customers");
+        parser.handleWayTags(intsRef, way, false, relFlags);
+        assertEquals(AccessRestrictionType.PRIVATE | AccessRestrictionType.CUSTOMERS, accessRestrictionEnc.getInt(false, intsRef));
     }
 
     @Test
@@ -91,7 +192,7 @@ class AccessRestrictionParserTest {
 
         way.setTag("bicycle", "destination");
         parser.handleWayTags(intsRef, way, false, relFlags);
-        assertEquals(AccessRestrictionType.NO | AccessRestrictionType.DESTINATION, accessRestrictionEnc.getInt(false, intsRef));
+        assertEquals(AccessRestrictionType.DESTINATION, accessRestrictionEnc.getInt(false, intsRef));
 
         way.setTag("bicycle", "yes");
         parser.handleWayTags(intsRef, way, false, relFlags);
@@ -109,7 +210,7 @@ class AccessRestrictionParserTest {
 
         way.setTag("foot", "destination");
         parser.handleWayTags(intsRef, way, false, relFlags);
-        assertEquals(AccessRestrictionType.NO | AccessRestrictionType.DESTINATION, accessRestrictionEnc.getInt(false, intsRef));
+        assertEquals(AccessRestrictionType.DESTINATION, accessRestrictionEnc.getInt(false, intsRef));
 
         way.setTag("foot", "yes");
         parser.handleWayTags(intsRef, way, false, relFlags);